+++ packages/patrol/lib/src/native/patrol_app_service.dart
@@ -148,6 +148,7 @@ class PatrolAppService extends PatrolAppServiceServer {
     _testExecutionRequested.complete(request.name);
 
     final testExecutionResult = await testExecutionCompleted;
+
     if (!testExecutionResult.passed) {
       _patrolLog.log(
         TestEntry(